Move error overlay middleware to server directory
